add simple test
